Remove the last vestiges of deleteEntry from AuctionServerManager.
git-svn-id: svn://svn.jbidwatcher.com/jbidwatcher/trunk@513 b1acfa68-eb39-11db-b167-a3a8cd6b847e

+++ src/com/jbidwatcher/ui/JBWDropHandler.java
@@ -63,10 +63,7 @@ public class JBWDropHandler implements MessageQueue.Listener {
         ErrorManagement.logDebug("Not loaded (" + aucId + ").");
         lastSeen = aucId;
       }
-      if(aeNew != null) {
-        AuctionServerManager.getInstance().deleteEntry(aeNew);
-        aeNew.delete();
-      }
+      if(aeNew != null) aeNew.delete();
     }
   }
